requirements.txt

argon2-cffi
bcrypt
flake8-docstrings===1.6.0
flake8===4.0.1
flask-mongoengine===1.0.0
flask-redis
mongoengine===0.24.1
mongopy===0.1
pendulum===2.1.2
pipenv-to-requirements
pipenv===2022.3.28
postgres
psycopg2
pytest===7.1.1
python-dateutil===2.8.2
python-decouple
python-dotenv===0.20.0
requests===2.27.1
wsproto===1.1.0
Django
Flask-BabelEx===0.9.4
Flask-Login===0.6.0
Flask-Mail===0.9.1
Flask-Principal===0.4.0
Flask-RBAC
Flask-Security===3.0.0
Flask-SQLAlchemy===2.5.1
Flask-WTF===1.0.1
Flask===2.0.3
MarkupSafe===2.0.1
SQLAlchemy===1.4.34
Werkzeug===2.0.0
WTForms-Alchemy
WTForms===3.0.1
Flask-AppBuilder
Jinja2===3.0.3